requirements.txt
praw
pymongo
spacy
python-dotenv